Complete Code Refactor
Split everything into files,
Fixed drag drop bug. Much more reliable now
Scroll bars also now fixed
