+++ VolumeMaster-ReactApp/src/assets/config.yaml
@@ -13,8 +13,8 @@ Mappings:
     ProcessNames: []
     VoiceMeeter: null
   "3":
-    ProcessNames:
+    ProcessNames: []
     VoiceMeeter: null
   "4":
-    ProcessNames:
+    ProcessNames: []
     VoiceMeeter: null
