+++ VolumeMaster-ReactApp/src/renderer/utils.js
@@ -1,0 +1,3 @@
+export function sanitizeAppName(name) {
+  return name.replace(/([A-Z]+)/g, ' $1').replace('.exe', '').trim();
+}
